Fix ; separated regexps: * add compatibility function for mac OS X missing getdelim * use delim for all cases in regexp2fst
git-svn-id: svn://svn.code.sf.net/p/hfst/code/trunk/hfst3@1233 941e2c2b-deac-454f-805a-451daa25f33c

+++ test/tools/regexp2fst-functionality.sh
@@ -1,7 +1,11 @@
 #!/bin/sh
 for i in sfst openfst-tropical foma; do
-if ! ../../tools/src/hfst-regexp2fst -f $i < cats_and_dogs.xre > test.hfst ; then
-    exit 1
-fi
-rm test.hfst;
+    if ! ../../tools/src/hfst-regexp2fst -f $i < cats_and_dogs.xre > test.hfst ; then
+        exit 1
+    fi
+    rm test.hfst;
+    if ! ../../tools/src/hfst-regexp2fst -S -f $i < cats_and_dogs_semicolon.xre > test.hfst ; then
+        exit 1
+    fi
+    rm test.hfst;
 done
